Fixed nDPI master/app brain damage.
When there was no app protocol detected, nDPI puts the master (real, ex: NetBIOS) protocol in the app_protocol field.  Make sense?  So now we check the value of the app_protocol, and if it less than the starting ID of our custom protocols, then we move it to master.  This works because we disable all "service" app protocols and define our own via app-custom-match.conf.