Added support for flow hash caching.
Every flow now has two hashes (digests).  A lower digest which is (as before) generated from fields from layer3 and below (protocol, addresses, ports, etc).  A "higher" or metadata digest which is generated from all available data (including layer3 and below).  These hashes are cached for a very long time (tunable), the more active a flow is, the "hotter" it becomes in the cache, living longer.  A cache search is performed on new flows and if the lower digest is found in the cache, the matching metadata hash is used as the flow digest (for consumption externally; netify-sink or via socket).  This ensures very long running flows can be "resurrected" instead of being classified as a new flow with DPI classification absent.